+++ lib/absinthe/schema/rule/object_must_implement_interfaces.ex
@@ -53,7 +53,7 @@ defmodule Absinthe.Schema.Rule.ObjectMustImplementInterfaces do
     |> Enum.map(&Schema.lookup_type(schema, &1))
     |> Enum.reduce([], fn
       %Type.Interface{} = iface_type, acc ->
-        if Type.Interface.implements?(iface_type, type) do
+        if Type.Interface.implements?(iface_type, type, schema) do
           acc
         else
           [report(type.__reference__.location, %{object: type.name, interface: iface_type.name}) | acc]